Move api code to api/v0
