+++ dokomoforms/handlers/api/v0/photos.py
@@ -1,5 +1,5 @@
 """TornadoResource class for dokomoforms.models.answer.Photo."""
-from dokomoforms.handlers.api import BaseResource
+from dokomoforms.handlers.api.v0 import BaseResource
 from dokomoforms.models import Photo, add_new_photo_to_session
