+++ dokomoforms/handlers/api/v0/submissions.py
@@ -6,7 +6,7 @@ from itertools import chain
 
 import restless.exceptions as exc
 
-from dokomoforms.handlers.api import BaseResource
+from dokomoforms.handlers.api.v0 import BaseResource
 from dokomoforms.models import (
     Survey, Submission, User,
     construct_submission, construct_answer, Answer,
